LayoutTests/http/tests/security/contentSecurityPolicy/frame-src-redirect-blocked-expected.txt

CONSOLE MESSAGE: Refused to frame 'http://localhost:8000/security/contentSecurityPolicy/resources/alert-fail.html' because it violates the following Content Security Policy directive: "frame-src 127.0.0.1:8000".
